Split mentionresponder from guild
